UnitConverter Unit Changed Changed the return value of class UnitConverter from Int to Float

+++ UnitConverter/UnitConverter.swift
@@ -9,8 +9,8 @@ import Foundation
 
 class UnitConverter {
     
-    func degreesFahrenheit(degreesCelsius: Int) -> Int {
-        return Int(1.8 * Float(degreesCelsius) + 32.0)
+    func degreesFahrenheit(degreesCelsius: Int) -> Float {
+        return Float(1.8 * Float(degreesCelsius) + 32.0)
     }
     
 }
